+++ web/src/router/index.js
@@ -4,7 +4,7 @@ import AdminLogin from '../views/Admin/AdminLogin.vue'
 import EmployeePage from '../views/Admin/EmployeePage.vue'
 import EmployeeMobile from '../views/Employee/EmployeeMobile.vue'
 import GeoMarkingList from '../components/GeoMarkingList.vue'
-import { getDeviceType  } from '../utlis/device.js'
+import { getDeviceType  } from '../utlis/general.js'
 import NotAllowed from '../components/NotAllowed.vue'
 import OwnerLoginPage from '../views/Owner/LoginPage.vue'
 import OwnerDashBoard from '../views/Owner/DashBoard.vue'
